requirements.txt
absl-py==1.4.0
anyio==3.6.2
asgiref==3.4.1
astor==0.8.1
backcall==0.2.0
bleach==1.5.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.0.4
cmake==3.28.4
colorama==0.4.5
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
entrypoints==0.4
enum34==1.1.10
fastapi==0.83.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h11==0.13.0
h5py==2.10.0
html5lib==0.9999999
idna==3.10
ImageHash==4.3.2
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mtcnn==0.1.1
nest-asyncio==1.6.0
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.9.2
Pygments==2.14.0
pyparsing==3.0.7
python-dateutil==2.9.0.post0
python-multipart==0.0.5
pytz==2025.2
PyWavelets==1.1.1
pywin32==305
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.17.0
sklearn==0.0
sniffio==1.2.0
starlette==0.19.1
tensorflow==1.15.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.16.0
xgboost==1.5.2
zipp==3.6.0
